backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.1
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
django-rql==4.4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
lark-parser==0.11.0
lib-rql==2.0.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
